requirements.txt
requests
json
pyyaml
django==1.8
